Last version emerging from our research team meeting of the 15th of january 2021
Modify the dataframe model. Split the project in several modules. Propose a new interface. Use a stylesheet to control its design. Manually test the behaviour and stability of the GUI when using other input file formats. This version was initially saved on the 19th of january 2021 and improved many times in march 2021.