Scale transformation
A scale transformation has been applied on 4 germination parameters : t1, t50, D, AUC, in order to render their statistical analysis comparable. But we observe that, eventually, the pvalues are exactly the same. So it's a useless transformation, the parametric tests for these 4 parameters are done with raw untransformed values. Other improvement : enlargement of scroll-bars.